Add eclipse.p2.wizardPageTitles property
Add new optional eclipse.p2.wizardPageTitles installer property that can
be used to set the titles for install wizard pages.  If the property is not
specified or a title for a page is not specified, the default page title
is used.
Change text of license page from "License Agreements" to
"Software Agreements".