Update
* Add support for specifying install wizard title area image using the
eclipse.p2.titleImage property.
* Add -filetemp option to INSTMON.  This is similar to the -file option that will
read command line options from a file, but will delete the file when done.
* Add -deleteOnExit option to INSTMON.  If this option is passed, INSTMON will
delete its executable after running.
* Updated INSTMON binaries in fragments.
* Fix race condition where P2 installer could delete options file passed to
INSTMON before INSTMON finished reading it.  The -filetemp option is now used
and INSTMON deletes the temporary options file after reading.
* Fix problem where running multiple uninstalls could result in a failure.  The
INSTMON binary is now copied to the system temporary directory instead of the
users home directory.  The -deleteOnExit option is used to delete INSTMON after
execution.
* Fix INSTMON log to be copied to the normal Eclipse configuration log directory
if it is not contained in the install area (-data).  Otherwise, the log is
copied to the installer data area in the home directory.
* Fix problem where old installs using previous version of path action would
fail to upgrade.
* Fix that if a non-existing IU is specified in the eclipse.p2.optionalRoots,
it is ignored and doesn't prevent silent installation.
* Fix install folder wizard page short-cut key for Browse... button.